Another CodeQL warning

+++ kitty/graphics.c
@@ -880,7 +880,7 @@ compose(const ComposeData d, uint8_t *under_data, const uint8_t *over_data) {
         uint8_t *under_row = under_data + (y + d.over_offset_y) * d.under_px_sz * d.under_width + d.under_px_sz * d.over_offset_x; \
         const uint8_t *over_row = over_data + y * d.over_px_sz * d.over_width;
     if (can_copy_rows) {
-        ROW_ITER memcpy(under_row, over_row, d.over_px_sz * min_row_sz);}
+        ROW_ITER memcpy(under_row, over_row, (size_t)d.over_px_sz * min_row_sz);}
         return;
     }
 #define PIX_ITER for (unsigned x = 0; x < min_row_sz; x++) { \
